+++ semcore/accordion/src/Accordion.jsx
@@ -142,7 +142,7 @@ class Toggle extends Component {
   };
 
   render() {
-    const { styles, disabled, use, Children } = this.asProps;
+    const { styles, disabled, use } = this.asProps;
     const SItemToggle = Root;
 
     return sstyled(styles)(
@@ -152,11 +152,7 @@ class Toggle extends Component {
         render={Text}
         onKeyDown={this.handleKeyDown}
         aria-disabled={disabled ? 'true' : undefined}
-      >
-        <Item.ToggleButton>
-          <Children />
-        </Item.ToggleButton>
-      </SItemToggle>,
+      />,
     );
   }
 }