+++ website/docs/components/accordion/examples/custom_styles.tsx
@@ -39,8 +39,10 @@ const Demo = () => {
                     selected && 'styled-accordion-item-selected',
                   )}
                 >
-                  <Accordion.Item.Chevron mr={2} />
-                  Section {index + 1}
+                  <Accordion.Item.ToggleButton>
+                    <Accordion.Item.Chevron mr={2} />
+                    Section {index + 1}
+                  </Accordion.Item.ToggleButton>
                 </Accordion.Item.Toggle>
                 <Accordion.Item.Collapse>
                   <Box p='12px 32px'>{`Hello Section ${index + 1}`}</Box>